project05_capstone/finished/requirements.txt
alabaster==0.7.10
alembic==1.2.0
astroid==1.5.3
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.5.0
backports.ssl==0.0.9
bcrypt==3.1.5
beautifulsoup4==4.5.3
bleach==2.0.0
blinker==1.4
botocore==1.13.44
certifi==2017.7.27.1
cffi==1.9.1
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==1.7.2
decorator==4.1.2
docutils==0.14
dominate==2.4.0
ecdsa==0.14.1
entrypoints==0.2.3
et-xmlfile==1.0.1
fabric==2.4.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-JWT-Simple==0.0.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Frozen-Flask==0.15
gunicorn==20.0.4
html5lib==0.999999999
idna==2.6
imagesize==0.7.1
IMAPClient==1.0.2
importlib-metadata==1.3.0
invoke==1.3.0
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
lxml==3.7.3
Mako==1.1.0
MarkupSafe==0.23
mccabe==0.6.1
mistune==0.7.4
mock==1.3.0
more-itertools==8.0.2
nbconvert==5.3.0
nbformat==4.4.0
nltk==3.2.4
numpy==1.13.1
numpydoc==0.7.0
oauthlib==2.0.2
olefile==0.44
openpyxl==2.4.4
packaging==19.2
pandocfilters==1.4.2
paramiko==2.6.0
pbr==2.0.0
pickleshare==0.7.4
Pillow==4.0.0
pluggy==0.13.1
prompt-toolkit==1.0.15
psutil==5.3.0
psycopg2==2.8.4
psycopg2-binary==2.8.3
py==1.8.0
pyasn1==0.2.3
PyAutoGUI==0.9.33
pycodestyle==2.3.1
pycparser==2.17
pyflakes==1.6.0
pygame==1.9.3
Pygments==2.2.0
PyJWT==1.4.2
pylint==1.7.2
pymongo==2.8
PyMsgBox==1.0.4
PyNaCl==1.3.0
pyOpenSSL==16.2.0
pyparsing==2.4.5
PyPDF2==1.26.0
pyperclip==1.5.27
PyScreeze==0.1.9
pytest==5.3.2
python-dateutil==2.6.0
python-docx==0.8.6
python-editor==1.0.4
python-jose==3.1.0
PyTweening==1.0.3
pytz==2017.2
PyYAML==5.2
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
requests-oauthlib==0.8.0
rope==0.10.7
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.19.1
selenium==3.0.2
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.2
SQLAlchemy==1.2.16
testpath==0.3.1
textblob==0.13.0
tornado==4.5.2
traitlets==4.3.2
tweepy==3.5.0
urllib3==1.22
virtualenv==16.7.7
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.0.0
xlwt==1.2.0
zipp==0.6.0
